ai_designer/backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# AI Models
torch==2.1.2
transformers==4.37.2
diffusers==0.25.1
accelerate==0.26.1
safetensors==0.4.1
pillow==10.2.0
opencv-python==4.9.0.80

# Gemini
google-generativeai==0.3.2

# CLIP
clip-by-openai==1.0
ftfy==6.1.3
regex==2023.12.25

# Image Processing
numpy==1.26.3
matplotlib==3.8.2
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Vector Database
qdrant-client==1.7.0

# Monitoring
prometheus-client==0.19.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
ruff==0.1.15
mypy==1.8.0
